refactor Entity from static singleton to EntityManager.cs and inject it in to MessageHandler.cs
